decom_function 메소드 명 decom_block으로 변경 +
decom_block에서 ir 데이터를 모아 IrFlowsInFunction생성하도록 적용
