+++ fireball/src/pe/fire/decom_from_virtual_address.rs
@@ -4,6 +4,6 @@ use crate::{core::Address, prelude::DecompileError};
 impl PE {
     pub(super) fn _decom_from_virtual_address(&self, address: u64) -> Result<(), DecompileError> {
         let address = Address::from_virtual_address(&self.sections, address);
-        self._decom_function(address)
+        self._decom_block(address)
     }
 }
